code review advices
