+++ src/App.tsx
@@ -5,14 +5,14 @@ import './App.scss';
 import posts from './api/posts';
 import comments from './api/comments';
 import users from './api/users';
-import { PostList } from './components/PostList';
+import { PostList } from './components/PostList/PostList';
 
-const preparedPosts = posts.map(key => {
-  const user = users.find(tmpUser => tmpUser.id === key.userId) || null;
-  const postComments = comments.filter(tmpComment => tmpComment.postId === key.id) || null;
+const preparedPosts = posts.map(post => {
+  const user = users.find(tmpUser => tmpUser.id === post.userId) || null;
+  const postComments = comments.filter(tmpComment => tmpComment.postId === post.id);
 
   return {
-    ...key,
+    ...post,
     user,
     comments: postComments,
   };
@@ -20,7 +20,7 @@ const preparedPosts = posts.map(key => {
 
 const App: React.FC = () => (
   <div className="App">
-    <h1>Static list of posts</h1>
+    <h1 className="title">Static list of posts</h1>
 
     <PostList preparedPosts={preparedPosts} />
   </div>
