--- src/components/CommentList.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import { Comment } from '../types/Comment';
-import { CommentInfo } from './CommentInfo';
-
-type Props = {
-  comments: Comment[]
-};
-
-export const CommentList: React.FC<Props> = ({ comments }) => (
-  <>
-    {comments.map(key => (
-      <CommentInfo comment={key} />
-    ))}
-  </>
-);
