+++ src/components/CommentList/CommentList.scss
@@ -1,0 +1,11 @@
+.comment {
+  display: flex;
+  flex-direction: column;
+  justify-content: end;
+  background-color: rgb(151, 221, 255);
+  border: 1px solid black;
+  border-radius: 25px;
+  margin: 5px 10px;
+  width: 70%;
+  padding: 10px;
+}
