+++ src/components/CommentList/CommentList.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { Comment } from '../../types/Comment';
+import { CommentInfo } from '../CommentInfo';
+import './CommentList.scss';
+
+type Props = {
+  comments: Comment[]
+};
+
+export const CommentList: React.FC<Props> = ({ comments }) => (
+  <>
+    {comments.map(comment => (
+      <div className="comment" key={comment.id}>
+        <CommentInfo comment={comment} />
+      </div>
+    ))}
+  </>
+);
