+++ src/components/PostInfo.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Post } from '../types/Post';
-import { UserInfo } from './UserInfo';
-import { CommentList } from './CommentList';
+import { UserInfo } from './UserInfo/UserInfo';
+import { CommentList } from './CommentList/CommentList';
 
 type Props = {
   post: Post
@@ -9,9 +9,11 @@ type Props = {
 
 export const PostInfo: React.FC<Props> = ({ post }) => (
   <>
-    <p>{post.title}</p>
+    <p className="title">{post.title}</p>
     <p>{post.body}</p>
     <UserInfo user={post.user} />
-    <CommentList comments={post.comments} />
+    <div className="comments">
+      <CommentList comments={post.comments} />
+    </div>
   </>
 );
