+++ src/components/PostList/PostList.scss
@@ -1,0 +1,15 @@
+.post {
+  background-color: #ddd;
+  border: 1px solid black;
+  border-radius: 10px;
+  margin-bottom: 30px;
+  margin-left: 10px;
+  margin-right: 10px;
+  padding: 10px;
+}
+
+.comments {
+  display: flex;
+  flex-direction: column;
+  align-items: end;
+}
