+++ src/components/UserInfo/UserInfo.scss
@@ -1,0 +1,5 @@
+.use-info {
+  text-align: right;
+  padding: 5px;
+  color: blue;
+}
